+++ packages/playground/src/front/js/components/IframeReloader.ts
@@ -1,7 +1,6 @@
 import { Base } from '@studiometa/js-toolkit';
 import type { BaseConfig, BaseProps } from '@studiometa/js-toolkit';
 
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
 export interface IframeReloaderProps extends BaseProps {}
 
 /**